Created separate folder for components and its related files
