+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 // import background from './bg.jpg'
-import Navbar from './components/navbar';
+import Navbar from './components/Navbar/navbar';
 
 function App() {
   return (
